Refactor FixedRadiusSearch (#3768)

+++ cpp/tests/core/NanoFlannIndex.cpp
@@ -60,7 +60,7 @@ TEST(NanoFlannIndex, SearchKnn) {
     core::Tensor distances;
     std::tie(indices, distances) = index.SearchKnn(query, 3);
 
-    ExpectEQ(indices.ToFlatVector<int64_t>(), std::vector<int64_t>({1, 4, 9}));
+    ExpectEQ(indices.ToFlatVector<int32_t>(), std::vector<int32_t>({1, 4, 9}));
     ExpectEQ(distances.ToFlatVector<double>(),
              std::vector<double>({0.00626358, 0.00747938, 0.0108912}));
     EXPECT_EQ(indices.GetShape(), core::SizeVector({1, 3}));
@@ -69,8 +69,8 @@ TEST(NanoFlannIndex, SearchKnn) {
     // if k > size
     std::tie(indices, distances) = index.SearchKnn(query, 12);
 
-    ExpectEQ(indices.ToFlatVector<int64_t>(),
-             std::vector<int64_t>({1, 4, 9, 0, 3, 2, 5, 7, 6, 8}));
+    ExpectEQ(indices.ToFlatVector<int32_t>(),
+             std::vector<int32_t>({1, 4, 9, 0, 3, 2, 5, 7, 6, 8}));
     ExpectEQ(distances.ToFlatVector<double>(),
              std::vector<double>({0.00626358, 0.00747938, 0.0108912, 0.0138322,
                                   0.015048, 0.018695, 0.0199108, 0.0286952,
